style: rename TuiSession.ts to kebab-case tui-session.ts
Aligns with the kebab-case convention used by other files in
src/tui-harness/lib/.
