+++ src/tui-harness/lib/tui-session.ts
